+++ NEWS.md
@@ -1,3 +1,5 @@
+# beadm (development version)
+
 # beadm v0.1.0
 
 `beadm` is a tool for managing ZFS boot environments on Linux. It is largely